+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/functions/hive/HiveSumAggFunction.java
@@ -106,6 +106,7 @@ public class HiveSumAggFunction extends HiveDeclarativeAggregateFunction {
     @Override
     public void setArguments(CallContext callContext) {
         if (resultType == null) {
+            checkArgumentCount(callContext.getArgumentDataTypes());
             resultType = initResultType(callContext.getArgumentDataTypes().get(0));
         }
     }
