+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/module/hive/HiveModule.java
@@ -88,7 +88,7 @@ public class HiveModule implements Module {
                                     "tumble_start")));
 
     static final Set<String> BUILTIN_NATIVE_AGG_FUNC =
-            Collections.unmodifiableSet(new HashSet<>(Arrays.asList("sum", "count", "min")));
+            Collections.unmodifiableSet(new HashSet<>(Arrays.asList("sum", "count", "min", "max")));
 
     private final HiveFunctionDefinitionFactory factory;
     private final String hiveVersion;
@@ -150,10 +150,6 @@ public class HiveModule implements Module {
         // We override some Hive's function by native implementation to supports hash-agg
         if (isNativeAggFunctionEnabled() && BUILTIN_NATIVE_AGG_FUNC.contains(name.toLowerCase())) {
             return getBuiltInNativeAggFunction(name.toLowerCase());
-        }
-
-        if (name.equalsIgnoreCase("max")) {
-            return Optional.of(new HiveMaxAggFunction());
         }
 
         // We override Hive's grouping function. Refer to the implementation for more details.
@@ -218,6 +214,9 @@ public class HiveModule implements Module {
             case "min":
                 // We override Hive's min function by native implementation to supports hash-agg
                 return Optional.of(new HiveMinAggFunction());
+            case "max":
+                // We override Hive's max function by native implementation to supports hash-agg
+                return Optional.of(new HiveMaxAggFunction());
             default:
                 throw new UnsupportedOperationException(
                         String.format(
